Device certificate generator encrpytion (#50)
Encryption of the returned certificates. #50 
According to the previous design, if the user does not provide a vault, instead of upload the certificates to a S3 bucket, the certificates would be passed back to the user.
This modified version requires the user to provide a key for AES encryption and the certificates would be encrypted.